add test for kb-h003
